backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rha02"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
python-multipart = "^0.0.12"
bcrypt = "^4.2.0"
python-dotenv = "^1.0.1"
pyjwt = "^2.9.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
azure-storage-blob = "^12.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"